updated some comments and renamed Profile page to Dashboard to make more sense
